requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0
datasets>=2.14.0

# Retrieval and Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.15

# Fine-tuning and Training
trl>=0.7.4
bitsandbytes>=0.41.0

# Evaluation and Analysis
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Safety and Content Filtering
detoxify>=0.5.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.14.0

# API and Deployment (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
